Update OddsPlotR.R.R
Added an option to customize the OR limits on the plot. This allows users to more easily standardize and compare ORs between graphs. 

The or_lim command must have a lower value specified to -1 to work. E.g., "or_lim=c(-1,20)" makes the limits of of the OR between 0 and 20.